+++ sklearn/covariance/_elliptic_envelope.py
